move dump to its own module
